Move WaitForChannel to the Opening stage
